Create sidebar component
