+++ src/ReceiptList.jsx
@@ -1,6 +1,9 @@
 import Receipt from './Receipt';
 
-const ReceiptList = ({ receipts }) =>
-  receipts.map((receipt) => <Receipt key={receipt.id} receipt={receipt} />);
+const ReceiptList = ({ receipts }) => {
+  return receipts.map((receipt) => (
+    <Receipt key={receipt.id} receipt={receipt} />
+  ));
+};
 
 export default ReceiptList;
